feat: rename project name "maeilmail"
