+++ Dockerfile
@@ -1,4 +1,4 @@
 FROM openjdk:17-jdk
-COPY build/libs/emailer-0.0.1-SNAPSHOT.jar emailer.jar
+COPY build/libs/maeilmail-0.0.1-SNAPSHOT.jar maeilmail.jar
 EXPOSE 8080
-ENTRYPOINT ["java", "-jar", "/emailer.jar"]
+ENTRYPOINT ["java", "-jar", "/maeilmail.jar"]
