+++ docker-compose.yml
@@ -2,7 +2,7 @@ version: '3.4'
 
 services:
   emailer:
-    image: emailer
+    image: maeilmail
     build:
       context: .
       dockerfile: ./Dockerfile
