+++ src/main/java/mongkey/maeilmail/MaeilmailApplication.java
@@ -1,13 +1,13 @@
-package team36.emailer;
+package mongkey.maeilmail;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class EmailerApplication {
+public class MaeilmailApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(EmailerApplication.class, args);
+        SpringApplication.run(MaeilmailApplication.class, args);
     }
 
 }
