+++ src/test/java/mongkey/maeilmail/MaeilmailApplicationTests.java
@@ -1,10 +1,10 @@
-package team36.emailer;
+package mongkey.maeilmail;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class EmailerApplicationTests {
+class MaeilmailApplicationTests {
 
     @Test
     void contextLoads() {
